Make external authentication optional
When configuration for external authentication (facebook, google, smartId,
mobileId) was omitted, the code would fail to start. This adds some checks and
skips initialization of these external authentication providers, allowing to
start without any of them.

Note that the these authentication providers should seperatelly be disabled in
the frontend config.

This diff is best viewed with `--ignore-all-space`.